Servant.OpenApi.Internal.TypeLevel.API

Documentation

type family EndpointsList api :: [Type] where ...

Build a list of endpoints from an API.

Equations

EndpointsList (a :<|> b) = AppendList (EndpointsList a) (EndpointsList b)

EndpointsList (e :> a) = MapSub e (EndpointsList a)

EndpointsList (NamedRoutes api) = EndpointsList (ToServantApi api)

EndpointsList a = '[a]

type family IsSubAPI sub api where ...

Check whether sub is a sub API of api.

Equations

IsSubAPI sub api = AllIsElem (EndpointsList sub) api

type family AllIsElem (xs :: [Type]) api where ...

Check that every element of xs is an endpoint of api.

Equations

AllIsElem ('[] :: [Type]) api = ()

AllIsElem (x ': xs) api = (IsIn x api, AllIsElem xs api)

type family MapSub (e :: k) (xs :: [Type]) :: [Type] where ...

Apply (e :>) to every API in xs.

Equations

MapSub (e :: k) ('[] :: [Type]) = '[] :: [Type]

MapSub (e :: k) (x ': xs) = (e :> x) ': MapSub e xs

type family AppendList (xs :: [a]) (ys :: [a]) :: [a] where ...

Append two type-level lists.

Equations

AppendList ('[] :: [a]) (ys :: [a]) = ys

AppendList (x ': xs :: [a]) (ys :: [a]) = x ': AppendList xs ys

type family Or a b where ...

Equations

Or () b = ()

Or a () = ()

type family IsIn sub api where ...

Equations

IsIn e (a :<|> b) = Or (IsIn e a) (IsIn e b)

IsIn (e :> a) (e :> b) = IsIn a b

IsIn e (NamedRoutes api) = IsIn e (ToServantApi api)

IsIn e e = ()

type family Elem (x :: a) (xs :: [a]) :: Bool where ...

Check whether a type is a member of a list of types. This is a type-level analogue of elem.

Equations

Elem (x :: a) ('[] :: [a]) = 'False

Elem (x :: a) (x ': xs :: [a]) = 'True

Elem (x :: a) (y ': xs :: [a]) = Elem x xs

type family Nub (xs :: [a]) :: [a] where ...

Remove duplicates from a type-level list.

Equations

Nub ('[] :: [a]) = '[] :: [a]

Nub (x ': xs :: [a]) = x ': Nub (Remove x xs)

type family Remove (x :: a) (xs :: [a]) :: [a] where ...

Remove element from a type-level list.

Equations

Remove (x :: a) ('[] :: [a]) = '[] :: [a]

Remove (x :: a) (x ': ys :: [a]) = Remove x ys

Remove (x :: a) (y ': ys :: [a]) = y ': Remove x ys

type BodyTypes c api = Nub (BodyTypes' c api)

Extract a list of unique "body" types for a specific content-type from a servant API.

type AddBodyType (c :: a1) (cs :: [a1]) (a2 :: a) (as :: [a]) = If (Elem c cs) (a2 ': as) as

AddBodyType c cs a as adds type a to the list as only if c is in cs.

type family BodyTypes' c api :: [Type] where ...

Extract a list of "body" types for a specific content-type from a servant API. To extract unique types see BodyTypes.

NoContent is removed from the list and not tested. (This allows for leaving the body completely empty on responses to requests that only accept 'application/json', while setting the content-type in the response accordingly.)

Equations

BodyTypes' c (Verb verb b cs (Headers hdrs a)) = AddBodyType c cs a ('[] :: [Type])

BodyTypes' c (Verb verb b cs NoContent) = '[] :: [Type]

BodyTypes' c (Verb verb b cs a) = AddBodyType c cs a ('[] :: [Type])

BodyTypes' c (ReqBody' mods cs a :> api) = AddBodyType c cs a (BodyTypes' c api)

BodyTypes' c (e :> api) = BodyTypes' c api

BodyTypes' c (a :<|> b) = AppendList (BodyTypes' c a) (BodyTypes' c b)

BodyTypes' c (NamedRoutes api) = BodyTypes' c (ToServantApi api)

BodyTypes' c api = '[] :: [Type]